Extract Advise proposal delivery client routes
