+++ quality/architecture_rules.md
@@ -35,11 +35,15 @@ report-only baseline is reviewed.
 
 The largest current modularity risks are:
 
-1. `src/app/clients/advise_proposal_client.py` at 406 lines,
-2. `src/app/contracts/proposal_lifecycle.py` at 405 lines,
-3. `src/app/services/proposal_service.py` at 405 lines,
-4. `src/app/services/platform_capabilities_normalization.py` at 404 lines,
-5. `src/app/services/performance_workspace_contribution.py` at 402 lines.
+1. `src/app/contracts/proposal_lifecycle.py` at 405 lines,
+2. `src/app/services/proposal_service.py` at 405 lines,
+3. `src/app/services/platform_capabilities_normalization.py` at 404 lines,
+4. `src/app/services/performance_workspace_contribution.py` at 402 lines,
+5. `src/app/services/risk_workspace_service.py` at 402 lines.
+
+`src/app/clients/advise_proposal_client.py` is reduced below the previous 406-line ceiling after
+proposal delivery, report-request, and execution route forwarding moved into
+`src/app/clients/advise_proposal_delivery_client.py`.
 
 `src/app/services/proposal_service.py` is reduced from 658 to 520 script-counted lines after
 proposal lifecycle transition orchestration moved into
